+++ spec/command/init_spec.rb
@@ -44,7 +44,7 @@ describe Conjur::Command::Init do
     it "returns the right certificate from github" do
       fingerprint, certificate = Conjur::Command::Init.get_certificate('github.com:443')
       expect(fingerprint).to eq(GITHUB_FP)
-      expect(certificate.strip).to eq(GITHUB_CERT.strip)
+      expect(certificate.strip).to include(GITHUB_CERT.strip)
     end
   end